Moved the pure uhd stuff around
