+++ CC/Sounder/BaseRadioSet-calibrate.cc
@@ -12,9 +12,7 @@
 */
 
 #include "include/BaseRadioSet.h"
-#include "include/BaseRadioSetUHD.h"
 #include "include/Radio.h"
-#include "include/RadioUHD.h"
 #include "include/comms-lib.h"
 #include "include/macros.h"
 #include "include/utils.h"
@@ -667,187 +665,4 @@ int BaseRadioSet::syncTimeOffset(bool measure_ref_radio, bool adjust_trigger) {
     }
   }
   return offset_diff;
-}
-
-int BaseRadioSetUHD::syncTimeOffsetUHD(bool measure_ref_radio, bool adjust_trigger) {
-    std::cout<<"sync time being called"<<std::endl;
-    int num_radios = _cfg->n_bs_sdrs().at(0) - 1;
-    if (num_radios < 1) {
-        std::cout << "No need to sample calibrate with one Iris! skipping ..."
-                  << std::endl;
-        return -1;
-    }
-    int seqLen = 160;  // Sequence length
-    std::vector<std::vector<float>> pilot =
-            CommsLib::getSequence(CommsLib::LTS_SEQ, seqLen);
-    auto pilot_cint16 = Utils::float_to_cint16(pilot);
-
-    // Prepend/Append vectors with prefix/postfix number of null samples
-    std::vector<std::complex<int16_t>> prefix_vec(_cfg->prefix(), 0);
-    std::vector<std::complex<int16_t>> postfix_vec(
-            _cfg->samps_per_slot() - _cfg->prefix() - seqLen, 0);
-    pilot_cint16.insert(pilot_cint16.begin(), prefix_vec.begin(),
-                        prefix_vec.end());
-    pilot_cint16.insert(pilot_cint16.end(), postfix_vec.begin(),
-                        postfix_vec.end());
-    // Transmitting from only one chain, create a null vector for chainB
-    std::vector<std::complex<int16_t>> dummy_cint16(pilot_cint16.size(), 0);
-    size_t num_samps = _cfg->samps_per_slot();
-
-    std::vector<void*> txbuff(2);
-    txbuff[0] = pilot_cint16.data();
-    if (_cfg->bs_sdr_ch() == 2) txbuff[1] = dummy_cint16.data();
-
-    std::vector<std::complex<int16_t>> dummyBuff0(num_samps);
-    std::vector<std::complex<int16_t>> dummyBuff1(num_samps);
-    std::vector<void*> dummybuffs(2);
-    dummybuffs[0] = dummyBuff0.data();
-    dummybuffs[1] = dummyBuff1.data();
-
-    long long txTime(0);
-    long long rxTime(0);
-
-    RadioUHD* ref_radio = bsRadios;
-    uhd::usrp::multi_usrp::sptr ref_dev = ref_radio->dev;
-
-    int offset_diff = num_samps;
-    if (measure_ref_radio == true) {
-        std::vector<std::complex<int16_t>> rx_buff_ref;
-        rx_buff_ref.resize(num_samps);
-
-        ref_radio->drain_buffers(dummybuffs, num_samps);
-        RadioUHD* front_radio = bsRadios;
-        uhd::usrp::multi_usrp::sptr front_dev = front_radio->dev;
-//        front_dev->setGain(SOAPY_SDR_TX, 0, "PAD", _cfg->cal_tx_gain().at(0));
-        front_dev->set_tx_gain(0,"PAD", _cfg->cal_tx_gain().at(0));
-        front_radio->activateXmit();
-        int ret = front_radio->xmit(txbuff.data(), num_samps, 3, txTime);
-        if (ret < 0) std::cout << "bad write" << std::endl;
-        ret = ref_radio->activateRecv(rxTime, num_samps, 3);
-        if (ret < 0) std::cout << "bad activate at front node" << std::endl;
-        radioTriggerUHD();
-        std::vector<void*> rxbuffref(2);
-        rxbuffref[0] = rx_buff_ref.data();
-        if (_cfg->bs_sdr_ch() == 2) rxbuffref[1] = dummyBuff0.data();
-        ret = ref_radio->recv(rxbuffref.data(), num_samps, rxTime);
-        auto rx = Utils::cint16_to_cfloat(rx_buff_ref);
-        //std::cout << "ref=[";
-        //for (size_t s = 0; s < num_samps; s++)
-        //  std::cout << rx_buff_ref.at(s).real() << "+1j*"
-        //            << rx_buff_ref.at(s).imag() << " ";
-        //std::cout << "];" << std::endl;
-        int peak = CommsLib::findLTS(rx, seqLen);
-        int offset = peak < seqLen ? 0 : peak - seqLen;
-        if (offset > 0) {
-            offset_diff = _cfg->prefix() - offset;
-            std::cout << "Ref offset from prefix is " << offset_diff << "."
-                      << std::endl;
-            /*std::cout << "adjusting delay of ref node by " << offset_diff
-                      << std::endl;*/
-            if (adjust_trigger) {
-                int abs_delta = std::abs(offset_diff);
-                if (abs_delta > _cfg->prefix())
-                    std::cout << "Likely bad receive!" << std::endl;
-                else {
-                    while (abs_delta > 0) {
-//                        ref_dev->writeSetting("ADJUST_DELAYS",
-//                                              offset_diff > 0 ? "1" : "-1");
-                        --abs_delta;
-                    }
-                }
-            }
-        }
-        front_radio->deactivateXmit();
-//        front_dev->setGain(SOAPY_SDR_TX, 0, "PAD", _cfg->tx_gain().at(0));
-        front_dev->set_tx_gain(0,"PAD", _cfg->cal_tx_gain().at(0));
-        ref_radio->deactivateRecv();
-        ref_radio->drain_buffers(dummybuffs, num_samps);
-    } else {
-        std::vector<std::vector<std::complex<int16_t>>> buff;
-        buff.resize(num_radios);
-        for (int i = 0; i < num_radios; i++) {
-            buff[i].resize(num_samps);
-        }
-
-//        ref_dev->setGain(SOAPY_SDR_TX, 0, "PAD", _cfg->cal_tx_gain().at(0));
-        ref_dev->set_tx_gain(0,"PAD", _cfg->cal_tx_gain().at(0));
-        for (int i = 0; i < num_radios; i++)
-            bsRadios->drain_buffers(dummybuffs, num_samps);
-
-        ref_radio->activateXmit();
-        int ret = ref_radio->xmit(txbuff.data(), num_samps, 3, txTime);
-        if (ret < 0) std::cout << "bad write" << std::endl;
-
-        // All write, or prepare to receive.
-        for (int j = 0; j < num_radios; j++) {
-            ret = bsRadios->activateRecv(rxTime, num_samps, 3);
-            if (ret < 0) std::cout << "bad activate at node " << j << std::endl;
-        }
-
-        radioTriggerUHD();
-
-        for (int i = 0; i < num_radios; i++) {
-            std::vector<void*> rxbuff(2);
-            rxbuff[0] = buff[i].data();
-            //rxbuff[1] = ant == 2 ? buff[(i*M+j)*ant+1].data() : dummyBuff.data();
-            if (_cfg->bs_sdr_ch() == 2) rxbuff[1] = dummyBuff0.data();
-            int ret = bsRadios->recv(rxbuff.data(), num_samps, rxTime);
-            if (ret < 0) std::cout << "bad read at node " << i << std::endl;
-        }
-
-        int min_offset = num_samps;
-        int max_offset = 0;
-        std::vector<int> offset(num_radios, 0);
-
-        for (int i = 0; i < num_radios; i++) {
-            // std::cout << "s" << i << "=[";
-            // for (size_t s = 0; s < num_samps; s++)
-            //     std::cout << buff[i].at(s).real() << "+1j*" << buff[i].at(s).imag()
-            //               << " ";
-            // std::cout << "];" << std::endl;
-            auto rx = Utils::cint16_to_cfloat(buff[i]);
-            int peak = CommsLib::findLTS(rx, seqLen);
-            offset[i] = peak < seqLen ? 0 : peak - seqLen;
-            //std::cout << i << " " << offset[i] << std::endl;
-            if (offset[i] != 0) {
-                min_offset = std::min(offset[i], min_offset);
-                max_offset = std::max(offset[i], max_offset);
-            }
-        }
-        std::cout << "Min/Max offset is " << min_offset << "/" << max_offset << "."
-                  << std::endl;
-        offset_diff = max_offset - min_offset;
-        if (std::abs(offset_diff) > kMaxOffsetDiff) {
-            std::cout << "Offset Diff is " << max_offset - min_offset
-                      << ". Likely bad receive!" << std::endl;
-        } else if (adjust_trigger && offset_diff > 1) {
-            // adjusting trigger delays based on lts peak index
-            std::vector<int> offset_sorted(offset.begin(), offset.end());
-            std::sort(offset_sorted.begin(), offset_sorted.end());
-            int median_offset = offset_sorted.at(num_radios / 2);
-            for (int i = 0; i < num_radios; i++) {
-                RadioUHD* bsRadio = bsRadios;
-                uhd::usrp::multi_usrp::sptr dev = bsRadio->dev;
-                //int delta = median_offset - offset[i];
-                int delta = median_offset - offset[i];
-                /*std::cout << "adjusting delay of node " << i << " by " << delta
-                          << std::endl;*/
-                int abs_delta = std::abs(delta);
-                while (abs_delta > 0) {
-//                    dev->writeSetting("ADJUST_DELAYS", delta > 0 ? "1" : "-1");
-                    --abs_delta;
-                }
-            }
-        }
-
-        ref_radio->deactivateXmit();
-//        ref_dev->setGain(SOAPY_SDR_TX, 0, "PAD", _cfg->tx_gain().at(0));
-        ref_dev->set_tx_gain(0,"PAD", _cfg->cal_tx_gain().at(0));
-        for (int i = 0; i < num_radios; i++) {
-            RadioUHD* bsRadio = bsRadios;
-            bsRadio->deactivateRecv();
-            bsRadio->drain_buffers(dummybuffs, num_samps);
-        }
-    }
-    return offset_diff;
 }
